+++ aimmo/views.py
@@ -13,21 +13,24 @@ from rest_framework.decorators import (
     api_view,
     authentication_classes,
     permission_classes,
+    action,
 )
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework.views import APIView
 
+from common.permissions import CanDeleteGame
+
 from . import game_renderer
+from .avatar_creator import create_avatar_for_user
 from .exceptions import UserCannotPlayGameException
-from .game_creator import create_avatar_for_user, create_game
 from .models import Avatar, Game
 from .permissions import (
     CanDeleteGameOrReadOnly,
     CsrfExemptSessionAuthentication,
     GameHasToken,
 )
-from .serializers import GameSerializer
+from .serializers import GameSerializer, GameIdsSerializer
 
 LOGGER = logging.getLogger(__name__)
 
@@ -50,7 +53,9 @@ def code(request, id):
         avatar.save()
         return HttpResponse(status=200)
     else:
-        return JsonResponse({"code": avatar.code})
+        return JsonResponse(
+            {"code": avatar.code, "starterCode": game.worksheet.starter_code}
+        )
 
 
 class GameUsersView(APIView):
@@ -88,6 +93,19 @@ class GameViewSet(
             serializer = GameSerializer(game)
             response[game.pk] = serializer.data
         return Response(response)
+
+    @action(
+        detail=False,
+        methods=["post"],
+        serializer_class=GameIdsSerializer,
+        permission_classes=(CanDeleteGame,),
+    )
+    def delete_games(self, request):
+        game_ids = request.data.getlist("game_ids")
+        Game.objects.filter(
+            pk__in=game_ids, game_class__teacher__new_user=request.user
+        ).delete()
+        return Response(status=status.HTTP_204_NO_CONTENT)
 
 
 @api_view(["GET"])