+++ game_frontend/src/redux/features/Editor/actions.js
@@ -1,37 +1,38 @@
 import types from './types'
 
 const getCodeRequest = () => ({
-  type: types.GET_CODE_REQUEST
+  type: types.GET_CODE_REQUEST,
 })
 
-const getCodeReceived = code => ({
+const getCodeReceived = (code, starterCode) => ({
   type: types.GET_CODE_SUCCESS,
   payload: {
-    code
-  }
+    code,
+    starterCode,
+  },
 })
 
-const postCodeRequest = code => ({
+const postCodeRequest = (code) => ({
   type: types.POST_CODE_REQUEST,
   payload: {
-    code
-  }
+    code,
+  },
 })
 
 const postCodeReceived = () => ({
-  type: types.POST_CODE_SUCCESS
+  type: types.POST_CODE_SUCCESS,
 })
 
 const avatarCodeUpdated = () => ({
-  type: types.AVATAR_CODE_UPDATED
+  type: types.AVATAR_CODE_UPDATED,
 })
 
 const resetCode = () => ({
-  type: types.RESET_CODE
+  type: types.RESET_CODE,
 })
 
 const codeReset = () => ({
-  type: types.CODE_RESET
+  type: types.CODE_RESET,
 })
 
 export default {
@@ -41,5 +42,5 @@ export default {
   postCodeReceived,
   avatarCodeUpdated,
   resetCode,
-  codeReset
+  codeReset,
 }